turn on circleci build

+++ devtools/ci/load_functions.sh
@@ -1,13 +1,7 @@
 #!/bin/sh
 
-url="http://ambermd.org/downloads/ambertools-dev/AmberTools18.tar.gz"
+url="http://ambermd.org/downloads/ambertools-dev/AmberTools18-dev.tar.gz"
 tarfile=`python -c "url='$url'; print(url.split('/')[-1])"`
-if [ "$CONDA" = "True" ]; then
-    ambertools_binary_url='https://480-81537431-gh.circle-artifacts.com/0/tmp/circle-artifacts.SHoEmLc/ambertools-build/amber-conda-bld/linux-64/ambertools-17.0-0.tar.bz2'
-else
-    ambertools_binary_url='https://480-81537431-gh.circle-artifacts.com/0/tmp/circle-artifacts.SHoEmLc/ambertools-build/amber-conda-bld/non-conda-install/linux-64.ambertools-17.0-0.16Apr17.H0417.tar.bz2'
-fi
-binary_tarfile=`python -c "url='${ambertools_binary_url}'; print(url.split('/')[-1])"`
 amber_version='16'
 ambertools_version='18'
 
@@ -25,31 +19,63 @@ function install_python(){
 
 
 function setup_ambertools(){
+    echo "HOME = $HOME"
     cwd=`pwd`
     mkdir $HOME/source_code
     cd $HOME/source_code
-    wget $url -O $tarfile
+    if [ ! -f $HOME/source_code/$tarfile ]; then
+        echo "Do not have $HOME/source_code/$tarfile"
+        echo "Downloading ..."
+        wget $url -O $tarfile
+    else
+        echo "Using cache folder $HOME/source_code/$tarfile"
+    fi
     tar -xf $tarfile
     install_python
 
     cd $HOME
-    wget ${ambertools_binary_url} -O ${binary_tarfile}
+    python $cwd/devtools/ci/download_circleci_AmberTools.py
     if [ "$CONDA" = "True" ]; then
+        binary_tarfile=`ls ambertools*${ambertools_version}*tar.bz2`
         conda install ${binary_tarfile}
     else
+        binary_tarfile=`ls linux-64.ambertools*${ambertools_version}*tar.bz2`
         tar -xf ${binary_tarfile}
     fi
     cd $cwd
 }
 
 
+function setup_ambertools_circleci(){
+    echo "Nothing"
+}
+
+
+make_test_links(){
+	cd $AMBERHOME/test/amd             && ln -s -f ../../bin/sander .
+	cd $AMBERHOME/test/chamber         && ln -s -f ../../bin/sander .
+	cd $AMBERHOME/test/nmropt          && ln -s -f ../../bin/sander .
+	cd $AMBERHOME/test/qmmm2           && ln -s -f ../../bin/sander .
+	cd $AMBERHOME/test/qmmm_DFTB       && ln -s -f ../../bin/sander .
+	cd $AMBERHOME/test/abfqmmm         && ln -s -f ../../bin/sander .
+	cd $AMBERHOME/test/rism3d          && ln -s -f ../../bin/sander .
+	cd $AMBERHOME/test/sander_pbsa_frc && ln -s -f ../../bin/sander .
+	cd $AMBERHOME/test/rism3d          && ln -s -f ../../bin/sander .
+}
+
+
 function run_tests(){
+    export PATH=$HOME/miniconda/bin:$PATH
     if [ "$CONDA" = "True" ]; then
         export AMBERHOME=`python -c "import sys; print(sys.prefix)"`
     else
         source $HOME/amber${ambertools_version}/amber.sh
     fi
-    amber.setup_test_folders $HOME/source_code/amber${amber_version}
-    python $TRAVIS_BUILD_DIR/devtools/ci/ci_test.py $TEST_TASK
-    # python $TRAVIS_BUILD_DIR/amber$version/AmberTools/src/conda_tools/amber.run_tests $TEST_TASK
+    $AMBERHOME/bin/amber.setup_test_folders $HOME/source_code/amber${amber_version}
+    make_test_links # FIXME: remove in amber?
+    git clone https://github.com/Amber-MD/ambertools-binary-build
+    cp ambertools-binary-build/conda_tools/amber.run_tests $AMBERHOME/bin/
+    git clone https://github.com/Amber-MD/ambertools-ci-base
+    $AMBERHOME/bin/amber.run_tests -t $TEST_TASK -x ambertools-ci-base/EXCLUDED_TESTS
+    cp test*.log $HOME/
 }
